+++ RELEASES.md
@@ -113,6 +113,7 @@ Upcoming
 - Add mime multipart response support to http sender
 - Wait indefinitely for adapter to start when job interval is 0
 - Add timeout field to IBIS console function 'Test a PipeLine'
+- Add 'multiple configs' checkbox field to Upload Configuration